Change behaviour of dblclick
This makes all surrouding bombs being revealed only when the user has
set a number grater than or equal to the number of surrouding bombs.

This is the behaviour of most games, when you're allowed to dig every
surrounding bomb, because if you flagged 3 squares when the game says
there is 3 squares in it, then you will take the chance and dig
everything surrounding it.