Put square-related functions into square.js
Remove dig function as it is really redundant. Call recursively to
reveal surrounding empty squares.

Use events to interact with the game:

For example, when a square is revealed, dispatch the event
`squareRevealed`. The callback should then reveal all bombs and end the
game if it was a bomb. Otherwise it should check if the player won.

Dispatch the event squareFlagged, to increment/decrement bombsCounter
conditionally on type of flag.