Add option to not dispatch squareRevealed event
This was causing a problem when a bomb was revealed and an
attempt to call revealAllBombs.

Not familiar with the error but this really should be prevented because
revealAllBombs would be called over and over, every time we revealed
another bomb.

When we fact we just want to show where they all are.