Add restart game functionality
"restart-game.js" module deals with the creation of a popup menu and
when it should be created, i.e., when the smiley face is clicked. The
popup is supposed to appear in the center of the game grid.

Also, it deals with the interface of the menu. When the user clicks yes,
we dispatch the event restartGame.

It also should appear and be hidden properly. It is hidden by default,
but appears when smiley is clicked. Hidden when one of the buttons is
clicked.

Finally, the game is restarted by calling the waitToStart method again,
but only if the user clicked the yes button.