Modify logic behind resetting game
Particularly, it is also considered changing difficulty as a game reset.

So a function askForNewGame was created, that "resets" the game
depending on its state.

If the game has started and isn't over, we ask for confirmation. If user
confirms (clicks yes butto), "restartGameConfirmed" event is triggered and
the game waits user to start playing.

If the game is over, just starts for user to start playing.

And if the game hasn't even started, we just change the grid.

So handleDifficultyChange method was removed. Changes in difficulty are
considered a NewGameRequest.