Add game submission mechanism
This is yet to be improved. Ideally we should ask for the user to enter
a nickname and then register all games, not just victories, under its
nickname into a database.

Currently this opens up a popup asking if he want to submit the game,
but this will change soon.