Dispatch squareRevealed event only on clicked square
This was causing trouble because when we clicked on an empty square, and
recursively called the function `reveal` (which also dispatches the
custom event squareRevealed), then the events would be fired several
times.

This caused trouble with the database because it would to try to write
the same game multiple times.

We only need to dispatch the event per square clicked, and then just
recursively call the `revealContent` function, which just change the
properties of the `<td>` element.