Fix bug of revealing surrounding squares with double click
When the user double clicked in a revealed square to reveal its
surroundings, a revealed bomb (e.g., if the user flagged a square in
which there was not a bomb) would not trigger all bombs to reveal (and
the game would continue...)

This commit just cleans the house to get rid of these problems.

I removed the auxiliary `revealContent`, whose functionality was to
manipulate the table data element attributes and change its appearance
(set its state to be revealed etc.). Now there is only one function
`reveal`.

If the square has a bomb, we set the boolean attribute 'bomb' in the
element and dispatch the event bombRevealed (which changes the smiley
content, records the game in the database and reveals all other bombs);

To reveal all bombs, I used the auxiliary function revealBomb, which
just set the html attribute 'bomb', because I don't want to trigger
again the event, which would retrigger the revealAllBombs method etc.

Now, if the square has no surrounding bombs, we just reveal them
recursively, without dispatching any event (maybe this is an error...,
but I'm not sure).

If the square has surrounding bombs, reveal that number and dispatch the
event to verify if the user won.