Add js script to sort table asc/desc when header is clicked
When the user clicks in a header, the column belonging to that header
will first be sorted in ascending order. Then this "target" will be
remembered and with additional clicks on that header will reverse the
order.

This also sets a attribute data-sort accordingly to display an arrow
besides the header text.

We also sort according to the column type. If the first item in the
column contains a '%', we will use a sort function that parses the float
first. If it is a number, we use the standard function to sort numbers
(subtraction). If it is a string, we use `localeCompare`.