Do not divide height equally per square, just the width
When width/height changes, divide the width of the game container
equally per square and, if this number is less than 25, apply them to
the squares styles.

This is by far the most frustrating thing in the app... I have no idea
how would be the best approach to resizing the square when the
width/height changes and still make the game playable.