Simplify start/finish logic of game
Use two functions, `resetGame` to reset the game state and `startGame` to
do the work needed for the game to start (place the bombs, start
tracking time etc.)

At every click, check if the game has started. If it hasn't, call
`startGame`. When the game is over or if a new game was confirmed we
reset the game.