Improve controls for touchscreen devices
Put a button on the game menu side wich toggles a click action. It is by
default the reveal action (bomb symbol). The user can click to change to
the flag action.

This is needed because in touch devices there is no right click so we
need to do everything with one event.

For more information on the isTouchScreen function, see
https://hacks.mozilla.org/2013/04/detecting-touch-its-the-why-not-the-how/.