Add width, height and bombs columns to leaderboards table
This is needed because of the random and custom difficulty, in which
case is important to give information about the board.

I might need to implement some filtering to this (I did, but the
possibilities/combinations of width and height is so enormous that would
be pointless), like filtering the displayed table (similar to the
sorting function that's already implemented.